Refactor type deduction
