+++ src/quill_type_deduce/src/lib.rs
@@ -1,6 +1,7 @@
+pub(crate) mod hindley_milner;
+pub mod hir;
 pub(crate) mod index_resolve;
 pub mod type_check;
-pub(crate) mod type_deduce;
 pub(crate) mod type_resolve;
 
 pub use index_resolve::replace_type_variables;
