+++ src/quill_type_deduce/src/type_check/mod.rs
@@ -1,15 +1,11 @@
 //! Performs type deduction and type checking of expressions and patterns.
 
-use std::{
-    collections::{hash_map::Entry, HashMap},
-    fmt::Display,
-};
+use std::collections::{hash_map::Entry, HashMap};
 
 use multimap::MultiMap;
 use quill_common::{
     diagnostic::{Diagnostic, DiagnosticResult, ErrorMessage, HelpMessage, HelpType, Severity},
     location::{Range, Ranged, SourceFileIdentifier},
-    name::QualifiedName,
 };
 use quill_index::{
     compute_used_files, DefinitionI, ProjectIndex, TypeDeclarationI, TypeDeclarationTypeI,
@@ -17,163 +13,23 @@ use quill_index::{
 };
 use quill_parser::{
     definition::{DefinitionBodyP, DefinitionCaseP},
-    expr_pat::{ConstantValue, ExprPatP},
+    expr_pat::ExprPatP,
     file::FileP,
     identifier::NameP,
 };
 use quill_type::{PrimitiveType, Type};
 
 use crate::{
-    index_resolve::{replace_type_variables, resolve_type_constructor, TypeConstructorInvocation},
-    type_deduce::deduce_expr_type,
+    hindley_milner::deduce_expr_type,
+    hir::{
+        definition::{Definition, DefinitionBody, DefinitionCase},
+        expr::{BoundVariable, Expression, TypeVariable},
+        pattern::Pattern,
+        SourceFileHIR,
+    },
+    index_resolve::{replace_type_variables, resolve_type_constructor},
     type_resolve::TypeVariableId,
 };
-
-/// A parsed and fully type checked source file.
-/// No effort has been made to ensure semantic consistency or correctness,
-/// just syntactic and type correctness.
-#[derive(Debug)]
-pub struct SourceFileHIR {
-    pub definitions: HashMap<String, Definition>,
-}
-
-impl Display for SourceFileHIR {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        writeln!(f, "Definitions:")?;
-        for (def_name, def) in &self.definitions {
-            writeln!(
-                f,
-                "  {} : {:?} -> {}",
-                def_name, def.arg_types, def.return_type
-            )?;
-        }
-        Ok(())
-    }
-}
-
-/// A definition for a symbol, i.e. a function or constant.
-/// The function's type is `arg_types -> return_type`.
-/// For example, if we defined a function
-/// ```notrust
-/// def foo: int -> int -> int {
-///     foo a b = a
-/// }
-/// ```
-/// then `arg_types` would be `[int, int]` and `return_type` would be `int`. If instead we defined it as
-/// ```notrust
-/// def foo: int -> int -> int {
-///     foo a = \b -> a
-/// }
-/// ```
-/// then `arg_types` would be `[int]` and `return_type` would be `int -> int`.
-#[derive(Debug)]
-pub struct Definition {
-    pub range: Range,
-    /// The type variables at the start of this definition.
-    pub type_variables: Vec<TypeParameter>,
-    pub arg_types: Vec<Type>,
-    pub return_type: Type,
-    pub body: DefinitionBody,
-}
-
-#[derive(Debug)]
-pub enum DefinitionBody {
-    PatternMatch(Vec<DefinitionCase>),
-    CompilerIntrinsic,
-}
-
-impl Ranged for Definition {
-    fn range(&self) -> Range {
-        self.range
-    }
-}
-
-#[derive(Debug)]
-pub struct DefinitionCase {
-    pub range: Range,
-    pub arg_patterns: Vec<Pattern>,
-    pub replacement: Expression,
-}
-
-/// A pattern made up of type constructors and potential unknowns.
-#[derive(Debug, Clone)]
-pub enum Pattern {
-    /// A name representing the entire pattern, e.g. `a`.
-    Named(NameP),
-    /// A constant value.
-    Constant { range: Range, value: ConstantValue },
-    /// A type constructor, e.g. `False` or `Maybe { value = a }`.
-    TypeConstructor {
-        type_ctor: TypeConstructorInvocation,
-        /// The list of fields. If a pattern is provided, the pattern is matched against the named field.
-        /// If no pattern is provided in Quill code, an automatic pattern is created, that simply assigns the field to a new variable with the same name.
-        fields: Vec<(NameP, Type, Pattern)>,
-    },
-    /// A function pattern. This cannot be used directly in code,
-    /// this is created only for working with functions that have multiple patterns.
-    Function {
-        param_types: Vec<Type>,
-        args: Vec<Pattern>,
-    },
-    /// An underscore representing an ignored pattern.
-    Unknown(Range),
-}
-
-impl Ranged for Pattern {
-    fn range(&self) -> Range {
-        match self {
-            Pattern::Named(identifier) => identifier.range,
-            Pattern::Constant { range, .. } => *range,
-            Pattern::TypeConstructor {
-                type_ctor,
-                fields: args,
-            } => args.iter().fold(type_ctor.range, |acc, (_name, _ty, pat)| {
-                acc.union(pat.range())
-            }),
-            Pattern::Unknown(range) => *range,
-            Pattern::Function { args, .. } => args
-                .iter()
-                .fold(args[0].range(), |acc, i| acc.union(i.range())),
-        }
-    }
-}
-
-impl Display for Pattern {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Pattern::Named(identifier) => write!(f, "{}", identifier.name),
-            Pattern::Constant { value, .. } => write!(f, "const {}", value),
-            Pattern::TypeConstructor {
-                type_ctor,
-                fields: args,
-            } => {
-                if args.is_empty() {
-                    return write!(f, "{}", type_ctor.data_type.name);
-                }
-
-                write!(f, "{} {{ ", type_ctor.data_type.name)?;
-                for (i, (name, _ty, pat)) in args.iter().enumerate() {
-                    if i != 0 {
-                        write!(f, ", ")?;
-                    }
-                    write!(f, " {}", name.name)?;
-                    write!(f, " = {}", pat)?;
-                }
-                write!(f, " }}")
-            }
-            Pattern::Function { args, .. } => {
-                for (i, arg) in args.iter().enumerate() {
-                    if i != 0 {
-                        write!(f, " ")?;
-                    }
-                    write!(f, "{}", arg)?;
-                }
-                Ok(())
-            }
-            Pattern::Unknown(_) => write!(f, "_"),
-        }
-    }
-}
 
 pub fn check(
     source_file: &SourceFileIdentifier,
@@ -193,61 +49,6 @@ struct TypeChecker<'a> {
     project_index: &'a ProjectIndex,
 
     messages: Vec<ErrorMessage>,
-}
-
-/// A variable bound by the definition of a function.
-#[derive(Debug, Clone)]
-pub struct BoundVariable {
-    pub range: Range,
-    pub var_type: Type,
-}
-
-/// A variable bound by some abstraction e.g. a lambda or a let inside it.
-#[derive(Debug, Clone)]
-pub struct AbstractionVariable {
-    pub range: Range,
-    pub var_type: TypeVariableId,
-}
-
-#[derive(Debug)]
-pub struct ExpressionT {
-    pub type_variable: TypeVariable,
-    pub contents: ExpressionContentsT,
-}
-
-impl Ranged for ExpressionT {
-    fn range(&self) -> Range {
-        self.contents.range()
-    }
-}
-
-/// Closely tied to the `Type` struct, this is used while type checking to allow
-/// unknown types, represented by `TypeVariableId`s.
-#[derive(Debug, Clone)]
-pub enum TypeVariable {
-    /// An explicitly named type possibly with type parameters, e.g. `Bool` or `Either[T, U]`.
-    Named {
-        name: QualifiedName,
-        parameters: Vec<TypeVariable>,
-    },
-    Function(Box<TypeVariable>, Box<TypeVariable>),
-    /// A known type variable, e.g. `T` or `F[A]`.
-    Variable {
-        variable: String,
-        parameters: Vec<TypeVariable>,
-    },
-    /// A completely unknown type - we don't even have a type variable letter to describe it such as `T`.
-    /// These are assigned random IDs, and when printed are converted into Greek letters using the
-    /// `TypeVariablePrinter`.
-    Unknown {
-        id: TypeVariableId,
-    },
-    /// A primitive type, built in to the core of the type system.
-    Primitive(PrimitiveType),
-    /// Borrow conditions are checked later.
-    Borrow {
-        ty: Box<TypeVariable>,
-    },
 }
 
 /// A utility for printing type variables to screen.
@@ -356,119 +157,6 @@ impl TypeVariablePrinter {
         }
     }
 }
-
-/// The Expression type is central to the HIR, or high-level intermediate representation.
-/// In an expression in HIR, the type of each object is known.
-#[derive(Debug)]
-pub struct Expression {
-    pub ty: Type,
-    pub contents: ExpressionContents,
-}
-
-impl Ranged for Expression {
-    fn range(&self) -> Range {
-        self.contents.range()
-    }
-}
-
-/// Represents the contents of an expression (which may or may not have been already type checked).
-/// The type `V` represents the type variables that we are substituting into this symbol.
-/// You should use `ExpressionContents` or `ExpressionContentsT` instead of this enum directly.
-#[derive(Debug)]
-pub enum ExpressionContentsGeneric<E, T, V> {
-    /// An argument to this function e.g. `x`.
-    Argument(NameP),
-    /// A local variable declared by a `lambda` or `let` expression.
-    Local(NameP),
-    /// A symbol in global scope e.g. `+` or `fold`.
-    Symbol {
-        /// The name that the symbol refers to.
-        name: QualifiedName,
-        /// The range where the symbol's name was written in this file.
-        range: Range,
-        /// The type variables we're substituting into this symbol.
-        /// If using an `ExpressionT`, this should be a vector of `TypeVariable`.
-        /// If using an `Expression`, this should be a vector of `Type`.
-        type_variables: V,
-    },
-    /// Apply the left hand side to the right hand side, e.g. `a b`.
-    /// More complicated expressions e.g. `a b c d` can be desugared into `((a b) c) d`.
-    Apply(Box<E>, Box<E>),
-    /// A lambda abstraction, specifically `lambda params -> expr`.
-    Lambda {
-        lambda_token: Range,
-        params: Vec<(NameP, T)>,
-        expr: Box<E>,
-    },
-    /// A let statement, specifically `let identifier = expr;`.
-    Let {
-        let_token: Range,
-        name: NameP,
-        expr: Box<E>,
-    },
-    /// A block of statements, inside parentheses, separated by line breaks or commas,
-    /// where the final expression in the block is the type of the block as a whole.
-    Block {
-        open_bracket: Range,
-        close_bracket: Range,
-        statements: Vec<E>,
-    },
-    /// Explicitly create a value of a data type.
-    ConstructData {
-        data_type_name: QualifiedName,
-        variant: Option<String>,
-        fields: Vec<(NameP, E)>,
-        open_brace: Range,
-        close_brace: Range,
-    },
-    /// A raw value, such as a string literal, the `unit` keyword, or an integer literal.
-    ConstantValue { value: ConstantValue, range: Range },
-    /// A borrowed value.
-    Borrow { borrow_token: Range, expr: Box<E> },
-    /// A copy of a borrowed value.
-    Copy { copy_token: Range, expr: Box<E> },
-}
-
-impl<E, T, V> Ranged for ExpressionContentsGeneric<E, T, V>
-where
-    E: Ranged,
-{
-    fn range(&self) -> Range {
-        match self {
-            ExpressionContentsGeneric::Argument(arg) => arg.range,
-            ExpressionContentsGeneric::Local(var) => var.range,
-            ExpressionContentsGeneric::Symbol { range, .. } => *range,
-            ExpressionContentsGeneric::Apply(l, r) => l.range().union(r.range()),
-            ExpressionContentsGeneric::Lambda {
-                lambda_token, expr, ..
-            } => lambda_token.union(expr.range()),
-            ExpressionContentsGeneric::Let {
-                let_token, expr, ..
-            } => let_token.union(expr.range()),
-            ExpressionContentsGeneric::ConstructData {
-                open_brace,
-                close_brace,
-                ..
-            } => open_brace.union(*close_brace),
-            ExpressionContentsGeneric::Block {
-                open_bracket,
-                close_bracket,
-                ..
-            } => open_bracket.union(*close_bracket),
-            ExpressionContentsGeneric::ConstantValue { range, .. } => *range,
-            ExpressionContentsGeneric::Borrow {
-                borrow_token, expr, ..
-            } => borrow_token.union(expr.range()),
-            ExpressionContentsGeneric::Copy {
-                copy_token, expr, ..
-            } => copy_token.union(expr.range()),
-        }
-    }
-}
-
-pub type ExpressionContents = ExpressionContentsGeneric<Expression, Type, Vec<Type>>;
-pub type ExpressionContentsT =
-    ExpressionContentsGeneric<ExpressionT, TypeVariable, HashMap<String, TypeVariable>>;
 
 /// Represents a declaration that may be in a different source file.
 pub struct ForeignDeclaration<T> {
